Add hint to Python-based JEXL parser and evaluator
When someone is considering using JEXL it might be important to note that there are implementations in other languages as well.

Especially helpful when the same expression needs to be evaluated in backend (which might be written e.g. Python) and JavaScript frontend. This makes JEXL even more useful than it already is.